convert to static components, use connect syntax
