+++ app/javascript/components/containers/StaticPage.js
@@ -6,6 +6,6 @@ import StaticPages from '../pages/static';
 
 export default class StaticPage extends PublicContainer {
   render() {
-    return <StaticPages { ...this.store.getState() } />;
+    return <StaticPages store={this.store} />;
   }
 }
