+++ app/javascript/components/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
 // Page reducers
-import site from '../pages/public-site/public-site-reducer';
+import site from '../redactions/site';
 
 const rootReducer = combineReducers({
   site
