+++ config/webpack/environment.js
@@ -1,3 +1,3 @@
 const { environment } = require('@rails/webpacker')
 
-module.exports = environment
+module.exports = environment;
